+++ packages/react-urlstack/react-urlstack-playground/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {StyleSheet, View, Text} from 'react-native'
-import {Navigator} from 'react-urlstack'
+import {Navigator} from '../react-urlstack'
 import routes from './routes'
-import SideMenu from './components/menu/SideMenu';
+import SideMenu from './components/menu/SideMenu'
 
 export default function StackApp(props){
   return (