+++ packages/react-urlstack/react-urlstack-playground/src/components/ScreenHeader.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import {StyleSheet, Text} from 'react-native';
-import {SharedElement} from 'react-urlstack';
+import {SharedElement} from '../../react-urlstack';
 
 function contentTransition( animatedValue, fromIndex, toIndex, box ){
 	let translation = animatedValue.interpolate({