+++ packages/react-urlstack/react-urlstack-playground/src/screens/Welcome.js
@@ -1,3 +1,12 @@
-import ContentComponent from '../components/ContentComponent'
+import React from 'react'
+import { View, Text, Button } from 'react-native'
 
-export default ContentComponent('Welcome to the react-urlstack playground.');
+export default function WelcomeScreen( props ){
+  console.log( props.drawer )
+  return (
+    <View>
+      <Text>Welcome to the playground</Text>
+      <Button onPress={ () => props.drawer.open() } title="Open drawer" />
+    </View>
+  )
+}